+++ test/integration/Hanji.t.sol
@@ -77,10 +77,6 @@ abstract contract HanjiTestBase is AllowanceHolderPairTest {
         return IHanjiPool(0xE27d2334Ab6402956c2E6E517d16fa206B3053ae);
     }
 
-    function uniswapV3Path() internal virtual override returns (bytes memory) {
-        return new bytes(0); // Not used for Hanji tests
-    }
-
     function uniswapV2Pool() internal virtual override returns (address) {
         return address(0); // Not used for Hanji tests
     }