+++ test/integration/SettlerPairTest.t.sol
@@ -18,6 +18,8 @@ import {Settler} from "src/Settler.sol";
 import {ISettlerActions} from "src/ISettlerActions.sol";
 import {RfqOrderSettlement} from "src/core/RfqOrderSettlement.sol";
 
+IERC20 constant wBTC = IERC20(0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599);
+
 abstract contract SettlerPairTest is SettlerBasePairTest {
     using SafeTransferLib for IERC20;
     using LibBytes for bytes;
@@ -60,7 +62,6 @@ abstract contract SettlerPairTest is SettlerBasePairTest {
         warmPermit2Nonce(MAKER);
     }
 
-    function uniswapV3Path() internal virtual returns (bytes memory);
     function uniswapV2Pool() internal virtual returns (address);
     function getCurveV2PoolData() internal pure virtual returns (ICurveV2Pool.CurveV2PoolData memory);
 
@@ -132,10 +133,10 @@ abstract contract SettlerPairTest is SettlerBasePairTest {
         snapEnd();
     }
 
-    function testSettler_uniswapV3VIP() public skipIf(uniswapV3Path().length == 0) {
+    function testSettler_uniswapV3VIP() public skipIf(uniswapV3PathVIP().length == 0) {
         (ISignatureTransfer.PermitTransferFrom memory permit, bytes memory sig) = _getDefaultFromPermit2();
         bytes[] memory actions = ActionDataBuilder.build(
-            abi.encodeCall(ISettlerActions.UNISWAPV3_VIP, (FROM, permit, uniswapV3Path(), sig, 0))
+            abi.encodeCall(ISettlerActions.UNISWAPV3_VIP, (FROM, permit, uniswapV3PathVIP(), sig, 0))
         );
         ISettlerBase.AllowedSlippage memory slippage = ISettlerBase.AllowedSlippage({
             recipient: payable(address(0)), buyToken: IERC20(address(0)), minAmountOut: 0 ether
@@ -166,6 +167,34 @@ abstract contract SettlerPairTest is SettlerBasePairTest {
             bytes32(0)
         );
         snapEnd();
+    }
+
+    function testSettler_uniswapV3VIP_multihop()
+        public
+        skipIf(uniswapV3PathVIP().length == 0)
+        skipIf(toToken() != WETH)
+    {
+        (ISignatureTransfer.PermitTransferFrom memory permit, bytes memory sig) = _getDefaultFromPermit2();
+        bytes memory path = uniswapV3PathVIP();
+        path = bytes.concat(path, abi.encodePacked(uint8(0), uint24(500), sqrtPriceLimitX96(toToken(), wBTC), wBTC));
+        bytes[] memory actions =
+            ActionDataBuilder.build(abi.encodeCall(ISettlerActions.UNISWAPV3_VIP, (FROM, permit, path, sig, 0)));
+        uint256 buyBalanceBefore = wBTC.balanceOf(FROM);
+        uint256 recipientIntermediateBalanceBefore = toToken().balanceOf(FROM);
+        uint256 settlerIntermediateBalanceBefore = toToken().balanceOf(address(settler));
+        ISettlerBase.AllowedSlippage memory slippage = ISettlerBase.AllowedSlippage({
+            recipient: payable(address(0)), buyToken: IERC20(address(0)), minAmountOut: 0 ether
+        });
+
+        Settler _settler = settler;
+        vm.startPrank(FROM);
+        snapStartName("settler_uniswapV3VIP_multihop");
+        _settler.execute(slippage, actions, bytes32(0));
+        snapEnd();
+
+        assertGt(wBTC.balanceOf(FROM), buyBalanceBefore);
+        assertEq(toToken().balanceOf(FROM), recipientIntermediateBalanceBefore);
+        assertEq(toToken().balanceOf(address(_settler)), settlerIntermediateBalanceBefore);
     }
 
     function testSettler_uniswapV3() public skipIf(uniswapV3Path().length == 0) {
@@ -214,10 +243,10 @@ abstract contract SettlerPairTest is SettlerBasePairTest {
         snapEnd();
     }
 
-    function testSettler_uniswapV3_buyToken_fee_single_custody() public skipIf(uniswapV3Path().length == 0) {
+    function testSettler_uniswapV3_buyToken_fee_single_custody() public skipIf(uniswapV3PathVIP().length == 0) {
         (ISignatureTransfer.PermitTransferFrom memory permit, bytes memory sig) = _getDefaultFromPermit2();
         bytes[] memory actions = ActionDataBuilder.build(
-            abi.encodeCall(ISettlerActions.UNISWAPV3_VIP, (address(settler), permit, uniswapV3Path(), sig, 0)),
+            abi.encodeCall(ISettlerActions.UNISWAPV3_VIP, (address(settler), permit, uniswapV3PathVIP(), sig, 0)),
             abi.encodeCall(
                 ISettlerActions.BASIC,
                 (
@@ -305,7 +334,6 @@ abstract contract SettlerPairTest is SettlerBasePairTest {
             defaultERC20PermitTransfer(address(fromToken()), amount(), PERMIT2_FROM_NONCE);
         bytes memory sig = getPermitTransferSignature(permit, address(settler), FROM_PRIVATE_KEY, permit2Domain);
         bytes memory permit2Action = abi.encodeCall(ISettlerActions.TRANSFER_FROM, (uniswapV2Pool(), permit, sig));
-        IERC20 wBTC = IERC20(0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599);
 
         // |7|6|5|4|3|2|1|0| - bit positions in swapInfo (uint8)
         // |0|0|0|0|0|0|F|Z| - Z: zeroForOne flag, F: sellTokenHasFee flag
@@ -369,7 +397,6 @@ abstract contract SettlerPairTest is SettlerBasePairTest {
     }
 
     function testSettler_uniswapV2_multihop() public skipIf(uniswapV2Pool() == address(0)) skipIf(toToken() != WETH) {
-        IERC20 wBTC = IERC20(0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599);
         address nextPool = 0xBb2b8038a1640196FbE3e38816F3e67Cba72D940; // UniswapV2 WETH/WBTC
 
         // |7|6|5|4|3|2|1|0| - bit positions in swapInfo (uint8)