+++ test/integration/USDT_WETH.t.sol
@@ -107,12 +107,7 @@ contract USDTWETHTest is
         */
     }
 
-    function uniswapV3Path()
-        internal
-        view
-        override(SettlerPairTest, AllowanceHolderPairTest, SettlerMetaTxnPairTest)
-        returns (bytes memory)
-    {
+    function uniswapV3Path() internal view override returns (bytes memory) {
         return abi.encodePacked(fromToken(), uint8(0), uint24(500), sqrtPriceLimitX96FromTo(), toToken());
     }